Added the features like login/siginup,Search different topics of discussion, add or delete comments by user

+++ base/views.py
@@ -1,7 +1,15 @@
-from django.shortcuts import render, redirect
 from django.http import HttpResponse
-from .models import Room
+from django.shortcuts import render, redirect
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room, Topic, Message
 from .forms import RoomForm
+import logging
 
 # Create your views here.
 
@@ -12,23 +20,107 @@ from .forms import RoomForm
 #    { 'id': 3 , 'name': "Frontend developers"},
    
    
-# ]
+# 
+
+
+
+def loginPage(request):
+
+    page = "login"
+
+    if request.user.is_authenticated:
+         return redirect("home")
+    if request.method == 'POST':
+        username = request.POST.get("username").lower()
+        password = request.POST.get('password')
+
+        
+
+        try:
+            user = User.objects.get(username__iexact=username)
+        except User.DoesNotExist:
+            messages.error(request, "User doesn't exist")
+            return redirect('login')  # Redirect to login page if user doesn't exist
+
+        user = authenticate(request, username=username, password=password)
+
+        # Debugging: Print authentication result
+        if user is not None:
+            print('User authenticated successfully.')
+            login(request, user)
+            return redirect('home')
+        else:
+            print('Authentication failed.')
+            messages.error(request, 'Username or password is incorrect')
+
+    context = {'page':page}
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+     logout(request)
+     return redirect('home')
+
+
+def registerPage(request):
+     
+     form = UserCreationForm()
+     if request.method == 'POST':
+          form = UserCreationForm(request.POST)
+
+          if form.is_valid():
+               user = form.save(commit=False)
+               user.username = user.username.lower()
+               user.save()
+               login(request, user)
+               return redirect ('home')
+        
+
+          else:
+               messages.error(request, "An error occured during registration")
+               
+
+          
+     return render (request, 'base/login_register.html', {'form': form})
 
 def home (request):
-    rooms = Room.objects.all() #gives all rooms from database
-    context = {'rooms': rooms}
+    q=request.GET.get('q') if request.GET.get("q")!= None else ''
+
+    
+    
+    rooms = Room.objects.filter(Q(topic__name__icontains = q) | Q(name__icontains=q)|  Q(description__icontains=q)) #gives all rooms from database
+    topics = Topic.objects.all()
+    room_count = rooms.count()
+
+    room_messages = Message.objects.all().filter(Q(room__name__icontains=q))
+
+    
+    context = {'rooms': rooms,'topics':topics, 'room_count':room_count, 'room_messages': room_messages}
     return render (request, "base/home.html", context)
 
 
 def room (request,pk):
         room = Room.objects.get(id=pk)
+
+        room_messages = room.message_set.all()
     
-        context = {'room': room}
+        participants = room.participants.all()
+
+        if request.method == 'POST':
+             message = Message.objects.create(
+                  user = request.user,
+                  room = room,
+                  body = request.POST.get('body')
+             )
+             room.participants.add(request.user)
+             return redirect ('room', pk=room.id)
+        context = {'room': room, "room_messages": room_messages,'participants':participants }
+
 
         return render (request, "base/room.html", context)
 
 
-
+@login_required(login_url= 'login')
 def createRoom(request):
       form = RoomForm()
       if request.method == "POST":
@@ -42,10 +134,13 @@ def createRoom(request):
       
       return render (request, 'base/room_form.html', context)
 
-
+@login_required(login_url= 'login')
 def updateRoom (request, pk):
       room = Room.objects.get(id=pk)
       form = RoomForm(instance = room)
+
+      if request.user!=room.host:
+           return HttpResponse("You are not allowed here !! ")
 
       if request.method== "POST":
             form = RoomForm(request.POST, instance = room)
@@ -55,12 +150,30 @@ def updateRoom (request, pk):
       context = {'form': form}
       return render (request, 'base/room_form.html', context)
 
-
+@login_required(login_url= 'login')
 def deleteRoom(request,pk):
       room = Room.objects.get(id=pk)
+
+      if request.user!=room.host:
+           return HttpResponse("You are not allowed here !! ")
 
       if request.method == "POST":
             room.delete()
             return redirect("home")
       return render (request, 'base/delete.html', {'obj': room})
 
+
+
+
+
+@login_required(login_url= 'login')
+def deleteMessage(request,pk):
+      message = Message.objects.get(id=pk)
+
+      if request.user!=message.user:
+           return HttpResponse("You are not allowed here !! ")
+
+      if request.method == "POST":
+            message.delete()
+            return redirect("home")
+      return render (request, 'base/delete.html', {'obj': message})
